Move `logger.py` to the top Revolve package
