+++ revolve/tol/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import
+
+from .logger import logger
+from .logger import output_console
+from .logger import log_debug
